Move Operation related helper functions to Operations class
