+++ hazelcast/src/main/java/com/hazelcast/spi/impl/operationservice/impl/OperationRunnerImpl.java
@@ -30,7 +30,7 @@ import com.hazelcast.partition.InternalPartition;
 import com.hazelcast.spi.BackupAwareOperation;
 import com.hazelcast.spi.Notifier;
 import com.hazelcast.spi.Operation;
-import com.hazelcast.spi.OperationAccessor;
+import com.hazelcast.spi.Operations;
 import com.hazelcast.spi.ReadonlyOperation;
 import com.hazelcast.spi.ResponseHandler;
 import com.hazelcast.spi.WaitSupport;
@@ -48,10 +48,10 @@ import com.hazelcast.util.ExceptionUtil;
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.logging.Level;
 
-import static com.hazelcast.spi.OperationAccessor.isJoinOperation;
-import static com.hazelcast.spi.OperationAccessor.isWanReplicationOperation;
 import static com.hazelcast.spi.OperationAccessor.setCallerAddress;
 import static com.hazelcast.spi.OperationAccessor.setConnection;
+import static com.hazelcast.spi.Operations.isJoinOperation;
+import static com.hazelcast.spi.Operations.isWanReplicationOperation;
 import static com.hazelcast.spi.impl.ResponseHandlerFactory.setRemoteResponseHandler;
 import static java.util.logging.Level.FINEST;
 import static java.util.logging.Level.SEVERE;
@@ -240,7 +240,7 @@ class OperationRunnerImpl extends OperationRunner {
     }
 
     private boolean retryDuringMigration(Operation op) {
-        return !(op instanceof ReadonlyOperation || OperationAccessor.isMigrationOperation(op));
+        return !(op instanceof ReadonlyOperation || Operations.isMigrationOperation(op));
     }
 
     private void handleOperationError(Operation operation, Throwable e) {
